Cleanup: Move not-so-generic components closer to use sites
While some of these components could easily qualify for a "shared
generic component" -status, they are not used in such a way at the
moment. Thus they belong with their respective use sites.
