+++ ui/src/components/common/index.ts
@@ -1,8 +1,5 @@
 export * from './BaseDateInput';
-export * from './DateControl';
 export * from './DateInput';
 export * from './FormInputTypes';
 export * from './PageTitle';
-export * from './RedirectWithQuery';
-export * from './TimeRangeControl';
 export * from './accordionClassNames';
