+++ ui/src/components/map/IconToggle.tsx
