+++ ui/src/components/routes-and-lines/search/RouteLineTableRow.tsx
@@ -10,16 +10,16 @@ import {
 import { isRoute } from '../../../graphql';
 import { Path, routeDetails } from '../../../router/routeDetails';
 import { MAX_DATE, MIN_DATE, mapToShortDate } from '../../../time';
-import { TimetableIcon } from '../../../uiComponents/TimetableIcon';
-import { VehicleIcon } from '../../../uiComponents/VehicleIcon';
-import { AlertPopover } from '../../common/AlertPopover';
 import { LocatorButton } from '../../common/Buttons';
 import { useAlertsAndHighLights } from '../../common/hooks';
 import { Column, Row, Visible } from '../../common/LayoutComponents';
 import { RouteLabel } from '../../common/RouteLabel';
+import { AlertPopover } from '../common/AlertPopover';
 import { LineDetailsButton } from './LineDetailsButton';
 import { LineTimetablesButton } from './LineTimetablesButton';
+import { TimetableIcon } from './TimetableIcon';
 import { RouteLineTableRowVariant } from './types';
+import { VehicleIcon } from './VehicleIcon';
 
 const testIds = {
   row: (testId: string) => `RouteLineTableRow::row::${testId}`,
