+++ ui/src/components/timetables/passing-times-by-stop/Popover.tsx
@@ -1,6 +1,6 @@
 import { FC, ReactNode } from 'react';
 import { twMerge } from 'tailwind-merge';
-import { CloseIconButton } from '../components/common/Buttons';
+import { CloseIconButton } from '../../common/Buttons';
 
 const testIds = {
   closeButton: 'Popover::closeButton',
