+++ ui/src/pages/CypressCoordinatesHelper.tsx
