+++ ui/src/router/Components/JoreLoader.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
-import { LoadingOverlay } from '../components/common/Loaders';
-import { useAppSelector } from '../hooks';
-import { selectIsJoreOperationLoading } from '../redux';
+import { LoadingOverlay } from '../../components/common/Loaders';
+import { useAppSelector } from '../../hooks';
+import { selectIsJoreOperationLoading } from '../../redux';
 
 const testIds = {
   loader: 'JoreLoader::loader',
