+++ ui/src/uiComponents/index.ts
@@ -1,5 +1,3 @@
 export * from './ConfirmationDialog';
 export * from './DialogWithButtons';
-export * from './IconToggle';
-export * from './Popover';
 export * from './Switch';
